update active link mark behavior
